src/requirements.txt
Django==2.1.7
djangorestframework==3.11.0
gunicorn==19.9.0
psycopg2-binary==2.8.6
PyJWT
drf-yasg==1.17.0
django-redis-cache==3.0.1
model_bakery
